🐛 fix : 디자인 및 Nav바 이름 수정, OK모달과 Fail 모달 합체
